Move UniqueToken to its own file
