+++ core/src/Kokkos_Threads.hpp
@@ -203,6 +203,7 @@ struct MemorySpaceAccess<Kokkos::Threads::memory_space,
 #include <Threads/Kokkos_Threads_Parallel_Range.hpp>
 #include <Threads/Kokkos_Threads_Parallel_MDRange.hpp>
 #include <Threads/Kokkos_Threads_Parallel_Team.hpp>
+#include <Threads/Kokkos_Threads_UniqueToken.hpp>
 
 #include <KokkosExp_MDRangePolicy.hpp>
